windows-app/requirements.txt
# Webcam Remota Universal - Dependências Python
# Interface gráfica
PyQt5==5.15.7
PyQt5-tools==5.15.4.3.2

# Processamento de vídeo e imagem
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1

# Áudio
pyaudio==0.2.11
pydub==0.25.1

# Rede e comunicação
websockets==11.0.3
requests==2.31.0
aiohttp==3.8.6

# Geração de executável
pyinstaller==6.0.0
auto-py-to-exe==2.34.0

# Utilitários
psutil==5.9.6
configparser==5.3.0